sharded: do not try to call nonexistent delete callback
If service object throws during creation its destructor may be called before
_delete_cb is set. Check for this.

+++ core/sharded.hh
@@ -39,7 +39,9 @@ class async_sharded_service : public enable_shared_from_this<T> {
 protected:
     std::function<void()> _delete_cb;
     ~async_sharded_service() {
-        _delete_cb();
+        if (_delete_cb) {
+            _delete_cb();
+        }
     }
     template <typename Service> friend class sharded;
 };
